requirements.txt
wagtail>=2.5,<2.6
wagtail-modeltranslation==0.10.2